requirements.txt
streamlit>=1.28.0
requests>=2.31.0
pillow>=10.0.0
opencv-python>=4.8.0
moviepy>=1.0.3
pydub>=0.25.1
openai>=1.0.0
websocket-client>=1.6.0
numpy>=1.24.0
psutil>=5.9.0